+++ solutions/devsprint-vinicius-carvalho-1/DeliveryAppChallenge/Models/DeliveryTime.swift
@@ -1,0 +1,18 @@
+//
+//  DeliveryTime.swift
+//  DeliveryAppChallenge
+//
+//  Created by Gustavo Soares on 25/01/22.
+//
+
+import Foundation
+
+struct DeliveryTime: Decodable {
+    let minimum: Int
+    let maximum: Int
+
+    private enum CodingKeys: String, CodingKey {
+        case minimum = "min"
+        case maximum = "max"
+    }
+}
