+++ solutions/devsprint-vinicius-carvalho-1/DeliveryAppChallenge/Models/Restaurant.swift
@@ -7,22 +7,6 @@
 
 import Foundation
 
-struct DeliveryTime: Decodable {
-    let minimum: Int
-    let maximum: Int
-
-    private enum CodingKeys: String, CodingKey {
-        case minimum = "min"
-        case maximum = "max"
-    }
-}
-
-struct Review: Decodable {
-    let score: Double
-    let count: Int
-}
-
-
 struct Restaurant: Decodable {
     let name: String
     let category: String
@@ -34,10 +18,4 @@ struct Restaurant: Decodable {
         case name, category, reviews, menu
         case deliveryTime = "delivery_time"
     }
-}
-
-struct RestaurantItem: Decodable {
-    let category: String
-    let name: String
-    let price: Double
 }
