+++ solutions/devsprint-vinicius-carvalho-1/DeliveryAppChallenge/Models/RestaurantItem.swift
@@ -1,0 +1,14 @@
+//
+//  RestaurantItem.swift
+//  DeliveryAppChallenge
+//
+//  Created by Gustavo Soares on 25/01/22.
+//
+
+import Foundation
+
+struct RestaurantItem: Decodable {
+    let category: String
+    let name: String
+    let price: Double
+}
