+++ solutions/devsprint-vinicius-carvalho-1/DeliveryAppChallenge/Models/Review.swift
@@ -1,0 +1,13 @@
+//
+//  Review.swift
+//  DeliveryAppChallenge
+//
+//  Created by Gustavo Soares on 25/01/22.
+//
+
+import Foundation
+
+struct Review: Decodable {
+    let score: Double
+    let count: Int
+}
